Fix Seaside continuation debugging with native code enabled and GemTools
There was an scenario in which it was not possible to debug Seaside
continuations from the object log. Most likely, this was when native
code was enabled and when trying to debug from GemTools. The issue was
related to the fact of using Breakpoints. After a large discussion with
Dale, we fixed this by using a regular "self halt" rather than
breakpoints.